refactor: RacingCarController 패키지 수정
